new pointer type for casting address in printf function, now supports 32 and 64 bit architecture

+++ lib/libft/printf_print.c
@@ -40,7 +40,7 @@ int	printf_p(void *ptr)
 
 	if (!ptr)
 		return (write(1, "(nil)", 5));
-	address = (unsigned long)ptr;
+	address = (uintptr_t)ptr;
 	printed = 0;
 	printed += write(1, "0x", 2);
 	ft_puthex(address, 0);
